requirements.txt
click>=8.1.2
notion-client>=2.0.0
python-dotenv>=0.17.0
vigor==0.1.1
